Fix gdb debug script auto-load via explicit data&debug configuration
The problem was that our default gdb datadir and debugdir had
non-standard paths, due to some unknown configuration detection.

In order to fix it, we now pass the standard paths via the --with-gdb-datadir & --with-separate-debug-dir variables.
I also set --with-jit-reader-dir, --with-system-gdbinit and --with-system-gdbinit-dir path to be the standard path as well.