build: Install gmp,mpfr,ncurses,expat
We now invoke make install for all of these packages. We later on add
their include paths so that the compiler is aware of any needed headers.

This allows us to not be reliant on *-dev libaries, which are often
libc-based, and are not installed in our custom musl toolchains.